+++ src/main/java/ch/heigvd/res/chill/domain/mhganh/LaCorne.java
@@ -1,13 +1,13 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.mhganh;
 
 import ch.heigvd.res.chill.domain.IProduct;
 
 import java.math.BigDecimal;
 
-public class PunkIPA implements IProduct {
+public class LaCorne implements IProduct {
 
-  public final static String NAME = "Punk IPA";
-  public final static BigDecimal PRICE = new BigDecimal(4.2);
+  public final static String NAME = "La Corne";
+  public final static BigDecimal PRICE = new BigDecimal(4.85);
 
   @Override
   public String getName() {
