+++ src/test/java/ch/heigvd/res/chill/domain/mhganh/LaFrancheTest.java
@@ -1,4 +1,4 @@
-package ch.heigvd.res.chill.domain.wasadigi;
+package ch.heigvd.res.chill.domain.mhganh;
 
 import ch.heigvd.res.chill.domain.Bartender;
 import ch.heigvd.res.chill.protocol.OrderRequest;
@@ -9,22 +9,22 @@ import java.math.BigDecimal;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-class PunkIPATest {
+class LaFrancheTest {
 
   @Test
-  void thePriceAndNameForPunkIPAShouldBeCorrect() {
-    PunkIPA beer = new PunkIPA();
-    assertEquals(beer.getName(), PunkIPA.NAME);
-    assertEquals(beer.getPrice(), PunkIPA.PRICE);
+  void thePriceAndNameForLaFrancheShouldBeCorrect() {
+    LaFranche beer = new LaFranche();
+    assertEquals(beer.getName(), LaFranche.NAME);
+    assertEquals(beer.getPrice(), LaFranche.PRICE);
   }
 
   @Test
-  void aBartenderShouldAcceptAnOrderForPunkIPA() {
+  void aBartenderShouldAcceptAnOrderForLaFranche() {
     Bartender jane = new Bartender();
-    String productName = "ch.heigvd.res.chill.domain.wasadigi.PunkIPA";
+    String productName = "ch.heigvd.res.chill.domain.mhganh.LaFranche";
     OrderRequest request = new OrderRequest(3, productName);
     OrderResponse response = jane.order(request);
-    BigDecimal expectedTotalPrice = PunkIPA.PRICE.multiply(new BigDecimal(3));
+    BigDecimal expectedTotalPrice = LaFranche.PRICE.multiply(new BigDecimal(3));
     assertEquals(expectedTotalPrice, response.getTotalPrice());
   }
